chore: 사용자 혼란 방지를 위해 컴포넌트명 교체 (#105)
