+++ src/app/(page)/dw2/page.tsx
@@ -1,6 +1,6 @@
 "use Client";
 
-import ColorPickerUI from "@components/ColorPicker/ColorPickerUI";
+import ColorPicker from "@components/ColorPicker/ColorPicker";
 import GradientPicker from "@components/ColorPicker/GradientPicker";
 import TwoPicker from "@components/ColorPicker/TwoPicker";
 
@@ -8,7 +8,7 @@ const dw2 = () => {
   return (
     <div className="mt-10 space-y-5">
       <TwoPicker />
-      <ColorPickerUI />
+      <ColorPicker />
       <GradientPicker />
     </div>
   );
